this is a more "clean" conversion to 64 bit than the type-fixups branch, which does borrow some elements from it but has far less find and replace...it is about 98-99% done, notes to follow later

+++ engine/server/library/serverGame/src/shared/commoditiesMarket/CommoditiesMarketListener.cpp
@@ -73,7 +73,7 @@ void CommoditiesMarketListener::OnAddAuction(int requestId, int result, const Ne
 			
 	}
 
-	CommoditiesMarket::onAddAuction(requestId, result, auctionOwnerId, ownerName, itemId, vendorId, vendorName, NetworkId(atoll(findNetworkId(location.c_str()))));
+	CommoditiesMarket::onAddAuction(requestId, result, auctionOwnerId, ownerName, itemId, vendorId, vendorName, NetworkId((int64_t) atoll(findNetworkId(location.c_str()))));
 }
 
 // ----------------------------------------------------------------------
@@ -93,7 +93,7 @@ void CommoditiesMarketListener::OnAddBid(int requestId, int result, const Networ
 //  toolow
 
 	
-	CommoditiesMarket::onAddBid(result, itemId, bidPlayerId, previousBidderId, maxProxyBidAmount, previousBid, NetworkId(atoll(findNetworkId(location.c_str()))), success, auctionOwnerId, ownerName, previousBidderName, itemNameLength, itemName, salesTaxAmount, salesTaxBankId);
+	CommoditiesMarket::onAddBid(result, itemId, bidPlayerId, previousBidderId, maxProxyBidAmount, previousBid, NetworkId((int64_t) atoll(findNetworkId(location.c_str()))), success, auctionOwnerId, ownerName, previousBidderName, itemNameLength, itemName, salesTaxAmount, salesTaxBankId);
 }
 
 // ----------------------------------------------------------------------
@@ -219,7 +219,7 @@ void CommoditiesMarketListener::OnCreateVendorMarket(int requestId, int result, 
 	UNREF(requestId); 
 	UNREF(ownerId);
 
-	CommoditiesMarket::onCreateVendorMarket(result, ownerId, NetworkId(atoll(findNetworkId(location.c_str()))));
+	CommoditiesMarket::onCreateVendorMarket(result, ownerId, NetworkId((int64_t) atoll(findNetworkId(location.c_str()))));
 }
 
 // ----------------------------------------------------------------------
